+++ src/com/madbros/tileminer/GameStates/LoadingState.java
@@ -25,20 +25,20 @@ public class LoadingState extends GameState{
 		Game.batch.setProjectionMatrix(Game.camera.combined);
 		Rect totalLoadingBar = new Rect(0,0,300, 40);
 		float percentage = Game.currentLoadingPoints/Game.totalLoadingPoints; //Giving me 0.3 if 3/10
-		float loadingBarWidth = totalLoadingBar.w*percentage;
+		float loadingBarWidth = (int)totalLoadingBar.w*percentage;
 		Rect loadingBar = new Rect(Game.currentScreenSizeX/2-(totalLoadingBar.w/2),Game.currentScreenSizeY/2-(totalLoadingBar.h/2),(int) loadingBarWidth, totalLoadingBar.h);
 		Game.batch.begin();
 			
 			//Rect test = new Rect(Game.currentScreenSizeX/2-150,Game.currentScreenSizeY/2-20, 300, 40);
 			
-			Sprites.pixel.draw(loadingBar.x-1,loadingBar.y-1,0,totalLoadingBar.w+2,totalLoadingBar.h+2);
+			Sprites.pixel.draw((int)loadingBar.x-1,(int)loadingBar.y-1,0,(int)totalLoadingBar.w+2,(int)totalLoadingBar.h+2);
 		
 			Sprites.pixel.setColor(Color.BLACK);
-			Sprites.pixel.draw(loadingBar.x,loadingBar.y,0,totalLoadingBar.w,totalLoadingBar.h);
+			Sprites.pixel.draw((int)loadingBar.x,(int)loadingBar.y,0,(int)totalLoadingBar.w,(int)totalLoadingBar.h);
 			Sprites.pixel.setColor(Color.WHITE);
 			
 			Sprites.pixel.setColor(Color.GREEN);
-			Sprites.pixel.draw(loadingBar.x,loadingBar.y,0,loadingBar.w,loadingBar.h);
+			Sprites.pixel.draw((int)loadingBar.x,(int)loadingBar.y,0,(int)loadingBar.w,(int)loadingBar.h);
 			Sprites.pixel.setColor(Color.WHITE);
 
 		Game.batch.end();
